redfin_scraper/requirements.txt
# Scrapy 및 기본 의존성
scrapy>=2.11.0
twisted>=22.10.0
cryptography>=41.0.0

# RSS 피드 파싱
feedparser>=6.0.10

# 날짜 파싱
python-dateutil>=2.8.2

# YAML 설정 파일
pyyaml>=6.0.1

# Pydantic (데이터 검증)
pydantic>=2.0.0

# HTTP 요청 및 응답 처리
requests>=2.31.0
urllib3>=2.0.0

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# 로깅 및 모니터링
loguru>=0.7.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 타입 힌트 지원
typing-extensions>=4.7.0

# 개발 도구
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pytest>=7.4.0
